requirements.txt
tensorflow-gpu==2.1.0
scikit-learn==0.22.1
opencv-python==4.2.0.32
editdistance
Pillow
